requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
httpx[http2]==0.28.1
jinja2==3.1.6
python-dotenv==1.2.2